Eliminated author.bsh; factored config handling to separate class.
