+++ import/index_java/src/org/solrmarc/index/VuFindIndexer.java
@@ -56,7 +56,6 @@ import org.solrmarc.callnum.LCCallNumber;
 import org.solrmarc.tools.CallNumUtils;
 import org.solrmarc.tools.SolrMarcIndexerException;
 import org.solrmarc.tools.Utils;
-import org.ini4j.Ini;
 
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
@@ -89,15 +88,8 @@ public class VuFindIndexer extends SolrIndexer
     private static final Pattern HDMSHDD_PATTERN = Pattern.compile("^([eEwWnNsS])(\\d+(\\.\\d+)?)");
     private static final Pattern PMDD_PATTERN = Pattern.compile("^([+-])(\\d+(\\.\\d+)?)");
 
-    private static ConcurrentHashMap<String, Ini> configCache = new ConcurrentHashMap<String, Ini>();
-    private ConcurrentHashMap<String, String> relatorSynonymLookup = new ConcurrentHashMap<String, String>();
-    private Set<String> knownRelators = new LinkedHashSet<String>();
-
     // Shutdown flag:
     private boolean shuttingDown = false;
-
-    // VuFind-specific configs:
-    private Properties vuFindConfigs = null;
 
     /**
      * Default constructor
@@ -109,11 +101,6 @@ public class VuFindIndexer extends SolrIndexer
     public VuFindIndexer(final String propertiesMapFile, final String[] propertyDirs)
             throws FileNotFoundException, IOException, ParseException {
         super(propertiesMapFile, propertyDirs);
-        try {
-            vuFindConfigs = Utils.loadProperties(propertyDirs, "vufind.properties");
-        } catch (IllegalArgumentException e) {
-            // If the properties load failed, don't worry about it -- we'll use defaults.
-        }
     }
 
     /**
@@ -127,166 +114,6 @@ public class VuFindIndexer extends SolrIndexer
     }
 
     /**
-     * Given the base name of a configuration file, locate the full path.
-     * @param filename base name of a configuration file
-     */
-    private File findConfigFile(String filename)
-    {
-        // Find VuFind's home directory in the environment; if it's not available,
-        // try using a relative path on the assumption that we are currently in
-        // VuFind's import subdirectory:
-        String vufindHome = System.getenv("VUFIND_HOME");
-        if (vufindHome == null) {
-            vufindHome = "..";
-        }
-
-        // Check for VuFind 2.0's local directory environment variable:
-        String vufindLocal = System.getenv("VUFIND_LOCAL_DIR");
-
-        // Get the relative VuFind path from the properties file, defaulting to
-        // the 2.0-style config/vufind if necessary.
-        String relativeConfigPath = Utils.getProperty(
-            vuFindConfigs, "vufind.config.relative_path", "config/vufind"
-        );
-
-        // Try several different locations for the file -- VuFind 2 local dir,
-        // VuFind 2 base dir, VuFind 1 base dir.
-        File file;
-        if (vufindLocal != null) {
-            file = new File(vufindLocal + "/" + relativeConfigPath + "/" + filename);
-            if (file.exists()) {
-                return file;
-            }
-        }
-        file = new File(vufindHome + "/" + relativeConfigPath + "/" + filename);
-        if (file.exists()) {
-            return file;
-        }
-        file = new File(vufindHome + "/web/conf/" + filename);
-        return file;
-    }
-
-    /**
-     * Sanitize a VuFind configuration setting.
-     * @param str configuration setting
-     */
-    private String sanitizeConfigSetting(String str)
-    {
-        // Drop comments if necessary:
-        int pos = str.indexOf(';');
-        if (pos >= 0) {
-            str = str.substring(0, pos).trim();
-        }
-
-        // Strip wrapping quotes if necessary (the ini reader won't do this for us):
-        if (str.startsWith("\"")) {
-            str = str.substring(1, str.length());
-        }
-        if (str.endsWith("\"")) {
-            str = str.substring(0, str.length() - 1);
-        }
-        return str;
-    }
-
-    /**
-     * Load an ini file.
-     * @param filename name of {@code .ini} file
-     */
-    public Ini loadConfigFile(String filename)
-    {
-        // Retrieve the file if it is not already cached.
-        if (!configCache.containsKey(filename)) {
-            Ini ini = new Ini();
-            try {
-                ini.load(new FileReader(findConfigFile(filename)));
-                configCache.putIfAbsent(filename, ini);
-            } catch (Throwable e) {
-                dieWithError("Unable to access " + filename);
-            }
-        }
-        return configCache.get(filename);
-    }
-
-    /**
-     * Get a section from a VuFind configuration file.
-     * @param filename configuration file name
-     * @param section section name within the file
-     */
-    public Map<String, String> getConfigSection(String filename, String section)
-    {
-        // Grab the ini file.
-        Ini ini = loadConfigFile(filename);
-        Map<String, String> retVal = ini.get(section);
-
-        String parent = ini.get("Parent_Config", "path");
-        while (parent != null) {
-            Ini parentIni = loadConfigFile(parent);
-            Map<String, String> parentSection = parentIni.get(section);
-            for (String key : parentSection.keySet()) {
-                if (!retVal.containsKey(key)) {
-                    retVal.put(key, parentSection.get(key));
-                }
-            }
-            parent = parentIni.get("Parent_Config", "path");
-        }
-
-        // Check to see if we need to worry about an override file:
-        String override = ini.get("Extra_Config", "local_overrides");
-        if (override != null) {
-            Map<String, String> overrideSection = loadConfigFile(override).get(section);
-            for (String key : overrideSection.keySet()) {
-                retVal.put(key, overrideSection.get(key));
-            }
-        }
-        return retVal;
-    }
-
-    /**
-     * Get a setting from a VuFind configuration file.
-     * @param filename configuration file name
-     * @param section section name within the file
-     * @param setting setting name within the section
-     */
-    public String getConfigSetting(String filename, String section, String setting)
-    {
-        String retVal = null;
-
-        // Grab the ini file.
-        Ini ini = loadConfigFile(filename);
-
-        // Check to see if we need to worry about an override file:
-        String override = ini.get("Extra_Config", "local_overrides");
-        if (override != null) {
-            Ini overrideIni = loadConfigFile(override);
-            retVal = overrideIni.get(section, setting);
-            if (retVal != null) {
-                return sanitizeConfigSetting(retVal);
-            }
-        }
-
-        // Try to find the requested setting:
-        retVal = ini.get(section, setting);
-
-        //  No setting?  Check for a parent configuration:
-        while (retVal == null) {
-            String parent = ini.get("Parent_Config", "path");
-            if (parent !=  null) {
-                try {
-                    ini.load(new FileReader(new File(parent)));
-                } catch (Throwable e) {
-                    dieWithError("Unable to access " + parent);
-                }
-                retVal = ini.get(section, setting);
-            } else {
-                break;
-            }
-        }
-
-        // Return the processed setting:
-        return retVal == null ? null : sanitizeConfigSetting(retVal);
-    }
-
-    /**
      * Connect to the VuFind database if we do not already have a connection.
      */
     private void connectToDatabase()
@@ -296,7 +123,7 @@ public class VuFindIndexer extends SolrIndexer
             return;
         }
 
-        String dsn = getConfigSetting("config.ini", "Database", "database");
+        String dsn = ConfigManager.instance().getConfigSetting("config.ini", "Database", "database");
 
         try {
             // Parse key settings from the PHP-style DSN:
@@ -1504,7 +1331,7 @@ public class VuFindIndexer extends SolrIndexer
      */
     public String[] getFulltextParserSettings()
     {
-        String parserType = getConfigSetting(
+        String parserType = ConfigManager.instance().getConfigSetting(
             "fulltext.ini", "General", "parser"
         );
         if (null != parserType) {
@@ -1512,7 +1339,7 @@ public class VuFindIndexer extends SolrIndexer
         }
 
         // Is Aperture active?
-        String aperturePath = getConfigSetting(
+        String aperturePath = ConfigManager.instance().getConfigSetting(
             "fulltext.ini", "Aperture", "webcrawler"
         );
         if ((null == parserType && null != aperturePath)
@@ -1523,7 +1350,7 @@ public class VuFindIndexer extends SolrIndexer
         }
 
         // Is Tika active?
-        String tikaPath = getConfigSetting(
+        String tikaPath = ConfigManager.instance().getConfigSetting(
             "fulltext.ini", "Tika", "path"
         );
         if ((null == parserType && null != tikaPath)
@@ -1763,684 +1590,6 @@ public class VuFindIndexer extends SolrIndexer
     public Logger getLogger()
     {
         return logger;
-    }
-
-    /**
-     * Extract all valid relator terms from a list of subfields using a whitelist.
-     * @param subfields        List of subfields to check
-     * @param permittedRoles   Whitelist to check against
-     * @param indexRawRelators Should we index relators raw, as found
-     * in the MARC (true) or index mapped versions (false)?
-     * @return Set of valid relator terms
-     */
-    public Set<String> getValidRelatorsFromSubfields(List<Subfield> subfields, List<String> permittedRoles, Boolean indexRawRelators)
-    {
-        Set<String> relators = new LinkedHashSet<String>();
-        for (int j = 0; j < subfields.size(); j++) {
-            String raw = subfields.get(j).getData();
-            String current = normalizeRelatorString(raw);
-            if (permittedRoles.contains(current)) {
-                relators.add(indexRawRelators ? raw : mapRelatorStringToCode(current));
-            }
-        }
-        return relators;
-    }
-
-    /**
-     * Is this relator term unknown to author-classification.ini?
-     * @param current relator to check
-     * @return True if unknown
-     */
-    public Boolean isUnknownRelator(String current)
-    {
-        // If we haven't loaded known relators yet, do so now:
-        if (knownRelators.size() == 0) {
-            Map<String, String> all = getConfigSection("author-classification.ini", "RelatorSynonyms");
-            for (String key : all.keySet()) {
-                knownRelators.add(normalizeRelatorString(key));
-                for (String synonym: all.get(key).split("\\|")) {
-                    knownRelators.add(normalizeRelatorString(synonym));
-                }
-            }
-        }
-        return !knownRelators.contains(normalizeRelatorString(current));
-    }
-
-    /**
-     * Extract all valid relator terms from a list of subfields using a whitelist.
-     * @param subfields      List of subfields to check
-     * @return Set of valid relator terms
-     */
-    public Set<String> getUnknownRelatorsFromSubfields(List<Subfield> subfields)
-    {
-        Set<String> relators = new LinkedHashSet<String>();
-        for (int j = 0; j < subfields.size(); j++) {
-            String current = subfields.get(j).getData().trim();
-            if (current.length() > 0 && isUnknownRelator(current)) {
-                logger.info("Unknown relator: " + current);
-                relators.add(current);
-            }
-        }
-        return relators;
-    }
-
-    /**
-     * Extract all values that meet the specified relator requirements.
-     * @param authorField           Field to analyze
-     * @param noRelatorAllowed      Array of tag names which are allowed to be used with
-     * no declared relator.
-     * @param relatorConfig         The setting in author-classification.ini which
-     * defines which relator terms are acceptable (or a colon-delimited list)
-     * @param unknownRelatorAllowed Array of tag names whose relators should be indexed 
-     * even if they are not listed in author-classification.ini.
-     * @param indexRawRelators      Set to "true" to index relators raw, as found
-     * in the MARC or "false" to index mapped versions.
-     * @return Set
-     */
-    public Set<String> getValidRelators(DataField authorField,
-        String[] noRelatorAllowed, String relatorConfig,
-        String[] unknownRelatorAllowed, String indexRawRelators
-    ) {
-        // get tag number from Field
-        String tag = authorField.getTag();
-        List<Subfield> subfieldE = authorField.getSubfields('e');
-        List<Subfield> subfield4 = authorField.getSubfields('4');
-
-        Set<String> relators = new LinkedHashSet<String>();
-
-        // if no relator is found, check to see if the current tag is in the "no
-        // relator allowed" list.
-        if (subfieldE.size() == 0 && subfield4.size() == 0) {
-            if (Arrays.asList(noRelatorAllowed).contains(tag)) {
-                relators.add("");
-            }
-        } else {
-            // If we got this far, we need to figure out what type of relation they have
-            List permittedRoles = normalizeRelatorStringList(Arrays.asList(loadRelatorConfig(relatorConfig)));
-            relators.addAll(getValidRelatorsFromSubfields(subfieldE, permittedRoles, indexRawRelators.toLowerCase().equals("true")));
-            relators.addAll(getValidRelatorsFromSubfields(subfield4, permittedRoles, indexRawRelators.toLowerCase().equals("true")));
-            if (Arrays.asList(unknownRelatorAllowed).contains(tag)) {
-                Set<String> unknown = getUnknownRelatorsFromSubfields(subfieldE);
-                if (unknown.size() == 0) {
-                    unknown = getUnknownRelatorsFromSubfields(subfield4);
-                }
-                relators.addAll(unknown);
-            }
-        }
-        return relators;
-    }
-
-    /**
-     * Parse a SolrMarc fieldspec into a map of tag name to set of subfield strings
-     * (note that we need to map to a set rather than a single string, because the
-     * same tag may repeat with different subfields to extract different sections
-     * of the same field into distinct values).
-     *
-     * @param tagList The field specification to parse
-     * @return HashMap
-     */
-    protected HashMap<String, Set<String>> getParsedTagList(String tagList)
-    {
-        String[] tags = tagList.split(":");//convert string input to array
-        HashMap<String, Set<String>> tagMap = new HashMap<String, Set<String>>();
-        //cut tags array up into key/value pairs in hash map
-        Set<String> currentSet;
-        for(int i = 0; i < tags.length; i++){
-            String tag = tags[i].substring(0, 3);
-            if (!tagMap.containsKey(tag)) {
-                currentSet = new LinkedHashSet<String>();
-                tagMap.put(tag, currentSet);
-            } else {
-                currentSet = tagMap.get(tag);
-            }
-            currentSet.add(tags[i].substring(3));
-        }
-        return tagMap;
-    }
-
-    /**
-     * Filter values retrieved using tagList to include only those whose relator
-     * values are acceptable. Used for separating different types of authors.
-     *
-     * @param record               The record (fed in automatically)
-     * @param tagList              The field specification to read
-     * @param acceptWithoutRelator Colon-delimited list of tags whose values should
-     * be accepted even if no relator subfield is defined
-     * @param relatorConfig        The setting in author-classification.ini which
-     * defines which relator terms are acceptable (or a colon-delimited list)
-     * @param acceptUnknownRelators Colon-delimited list of tags whose relators
-     * should be indexed even if they are not listed in author-classification.ini.
-     * @param indexRawRelators      Set to "true" to index relators raw, as found
-     * in the MARC or "false" to index mapped versions.
-     * @param firstOnly            Return first result only?
-     * @return List result
-     */
-    public List<String> getAuthorsFilteredByRelator(Record record, String tagList,
-        String acceptWithoutRelator, String relatorConfig,
-        String acceptUnknownRelators, String indexRawRelators, Boolean firstOnly
-    ) {
-        List<String> result = new LinkedList<String>();
-        String[] noRelatorAllowed = acceptWithoutRelator.split(":");
-        String[] unknownRelatorAllowed = acceptUnknownRelators.split(":");
-        HashMap<String, Set<String>> parsedTagList = getParsedTagList(tagList);
-        List fields = getFieldSetMatchingTagList(record, tagList);
-        Iterator fieldsIter = fields.iterator();
-        if (fields != null){
-            DataField authorField;
-            while (fieldsIter.hasNext()){
-                authorField = (DataField) fieldsIter.next();
-                // add all author types to the result set; if we have multiple relators, repeat the authors
-                for (String iterator: getValidRelators(authorField, noRelatorAllowed, relatorConfig, unknownRelatorAllowed, indexRawRelators)) {
-                    for (String subfields : parsedTagList.get(authorField.getTag())) {
-                        String current = getDataFromVariableField(authorField, "["+subfields+"]", " ", false);
-                        // TODO: we may eventually be able to use this line instead,
-                        // but right now it's not handling separation between the
-                        // subfields correctly, so it's commented out until that is
-                        // fixed.
-                        //String current = authorField.getSubfieldsAsString(subfields);
-                        if (null != current) {
-                            result.add(current);
-                            if (firstOnly) {
-                                return result;
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        return result;
-    }
-
-    /**
-     * Filter values retrieved using tagList to include only those whose relator
-     * values are acceptable. Used for separating different types of authors.
-     *
-     * @param record               The record (fed in automatically)
-     * @param tagList              The field specification to read
-     * @param acceptWithoutRelator Colon-delimited list of tags whose values should
-     * be accepted even if no relator subfield is defined
-     * @param relatorConfig        The setting in author-classification.ini which
-     * defines which relator terms are acceptable (or a colon-delimited list)
-     * @return List result
-     */
-    public List<String> getAuthorsFilteredByRelator(Record record, String tagList,
-        String acceptWithoutRelator, String relatorConfig
-    ) {
-        // default firstOnly to false!
-        return getAuthorsFilteredByRelator(
-            record, tagList, acceptWithoutRelator, relatorConfig,
-            acceptWithoutRelator, "false", false
-        );
-    }
-
-    /**
-     * Filter values retrieved using tagList to include only those whose relator
-     * values are acceptable. Used for separating different types of authors.
-     *
-     * @param record               The record (fed in automatically)
-     * @param tagList              The field specification to read
-     * @param acceptWithoutRelator Colon-delimited list of tags whose values should
-     * be accepted even if no relator subfield is defined
-     * @param relatorConfig        The setting in author-classification.ini which
-     * defines which relator terms are acceptable (or a colon-delimited list)
-     * @return List result
-     * @param acceptUnknownRelators Colon-delimited list of tags whose relators
-     * should be indexed even if they are not listed in author-classification.ini.
-     */
-    public List<String> getAuthorsFilteredByRelator(Record record, String tagList,
-        String acceptWithoutRelator, String relatorConfig,
-        String acceptUnknownRelators
-    ) {
-        // default firstOnly to false!
-        return getAuthorsFilteredByRelator(
-            record, tagList, acceptWithoutRelator, relatorConfig,
-            acceptUnknownRelators, "false", false
-        );
-    }
-
-    /**
-     * Filter values retrieved using tagList to include only those whose relator
-     * values are acceptable. Used for separating different types of authors.
-     *
-     * @param record               The record (fed in automatically)
-     * @param tagList              The field specification to read
-     * @param acceptWithoutRelator Colon-delimited list of tags whose values should
-     * be accepted even if no relator subfield is defined
-     * @param relatorConfig        The setting in author-classification.ini which
-     * defines which relator terms are acceptable (or a colon-delimited list)
-     * @return List result
-     * @param acceptUnknownRelators Colon-delimited list of tags whose relators
-     * should be indexed even if they are not listed in author-classification.ini.
-     * @param indexRawRelators      Set to "true" to index relators raw, as found
-     * in the MARC or "false" to index mapped versions.
-     */
-    public List<String> getAuthorsFilteredByRelator(Record record, String tagList,
-        String acceptWithoutRelator, String relatorConfig,
-        String acceptUnknownRelators, String indexRawRelators
-    ) {
-        // default firstOnly to false!
-        return getAuthorsFilteredByRelator(
-            record, tagList, acceptWithoutRelator, relatorConfig,
-            acceptUnknownRelators, indexRawRelators, false
-        );
-    }
-
-    /**
-     * If the provided relator is included in the synonym list, convert it back to
-     * a code (for better standardization/translation).
-     *
-     * @param relator Relator code to check
-     * @return Code version, if found, or raw string if no match found.
-     */
-    public String mapRelatorStringToCode(String relator)
-    {
-        String normalizedRelator = normalizeRelatorString(relator);
-        return relatorSynonymLookup.containsKey(normalizedRelator)
-            ? relatorSynonymLookup.get(normalizedRelator) : relator;
-    }
-
-    /**
-     * Filter values retrieved using tagList to include only those whose relator
-     * values are acceptable. Used for separating different types of authors.
-     *
-     * @param record                The record (fed in automatically)
-     * @param tagList               The field specification to read
-     * @param acceptWithoutRelator  Colon-delimited list of tags whose values should
-     * be accepted even if no relator subfield is defined
-     * @param relatorConfig         The setting in author-classification.ini which
-     * defines which relator terms  are acceptable (or a colon-delimited list)
-     * @param acceptUnknownRelators Colon-delimited list of tags whose relators
-     * should be indexed even if they are not listed in author-classification.ini.
-     * @param indexRawRelators      Set to "true" to index relators raw, as found
-     * in the MARC or "false" to index mapped versions.
-     * @return String
-     */
-    public String getFirstAuthorFilteredByRelator(Record record, String tagList,
-        String acceptWithoutRelator, String relatorConfig,
-        String acceptUnknownRelators, String indexRawRelators
-    ) {
-        List<String> result = getAuthorsFilteredByRelator(
-            record, tagList, acceptWithoutRelator, relatorConfig,
-            acceptUnknownRelators, indexRawRelators, true
-        );
-        for (String s : result) {
-            return s;
-        }
-        return null;
-    }
-
-    /**
-     * Filter values retrieved using tagList to include only those whose relator
-     * values are acceptable. Used for separating different types of authors.
-     *
-     * @param record                The record (fed in automatically)
-     * @param tagList               The field specification to read
-     * @param acceptWithoutRelator  Colon-delimited list of tags whose values should
-     * be accepted even if no relator subfield is defined
-     * @param relatorConfig         The setting in author-classification.ini which
-     * defines which relator terms  are acceptable (or a colon-delimited list)
-     * @return String
-     */
-    public String getFirstAuthorFilteredByRelator(Record record, String tagList,
-        String acceptWithoutRelator, String relatorConfig
-    ) {
-        return getFirstAuthorFilteredByRelator(
-            record, tagList, acceptWithoutRelator, relatorConfig,
-            acceptWithoutRelator, "false"
-        );
-    }
-
-    /**
-     * Filter values retrieved using tagList to include only those whose relator
-     * values are acceptable. Used for separating different types of authors.
-     *
-     * @param record                The record (fed in automatically)
-     * @param tagList               The field specification to read
-     * @param acceptWithoutRelator  Colon-delimited list of tags whose values should
-     * be accepted even if no relator subfield is defined
-     * @param relatorConfig         The setting in author-classification.ini which
-     * defines which relator terms  are acceptable (or a colon-delimited list)
-     * @param acceptUnknownRelators Colon-delimited list of tags whose relators
-     * should be indexed even if they are not listed in author-classification.ini.
-     * @return String
-     */
-    public String getFirstAuthorFilteredByRelator(Record record, String tagList,
-        String acceptWithoutRelator, String relatorConfig,
-        String acceptUnknownRelators
-    ) {
-        return getFirstAuthorFilteredByRelator(
-            record, tagList, acceptWithoutRelator, relatorConfig,
-            acceptUnknownRelators, "false"
-        );
-    }
-
-    /**
-     * Filter values retrieved using tagList to include only those whose relator
-     * values are acceptable. Used for saving relators of authors separated by different
-     * types.
-     *
-     * @param record               The record (fed in automatically)
-     * @param tagList              The field specification to read
-     * @param acceptWithoutRelator Colon-delimited list of tags whose values should
-     * be accepted even if no relator subfield is defined
-     * @param relatorConfig        The setting in author-classification.ini which
-     * defines which relator terms are acceptable (or a colon-delimited list)
-     * @param acceptUnknownRelators Colon-delimited list of tags whose relators
-     * should be indexed even if they are not listed in author-classification.ini.
-     * @param indexRawRelators      Set to "true" to index relators raw, as found
-     * in the MARC or "false" to index mapped versions.
-     * @param firstOnly            Return first result only?
-     * @return List result
-     */
-    public List getRelatorsFilteredByRelator(Record record, String tagList,
-        String acceptWithoutRelator, String relatorConfig,
-        String acceptUnknownRelators, String indexRawRelators, Boolean firstOnly
-    ) {
-        List result = new LinkedList();
-        String[] noRelatorAllowed = acceptWithoutRelator.split(":");
-        String[] unknownRelatorAllowed = acceptUnknownRelators.split(":");
-        HashMap<String, Set<String>> parsedTagList = getParsedTagList(tagList);
-        List fields = getFieldSetMatchingTagList(record, tagList);
-        Iterator fieldsIter = fields.iterator();
-        if (fields != null){
-            DataField authorField;
-            while (fieldsIter.hasNext()){
-                authorField = (DataField) fieldsIter.next();
-                //add all author types to the result set
-                result.addAll(getValidRelators(authorField, noRelatorAllowed, relatorConfig, unknownRelatorAllowed, indexRawRelators));
-            }
-        }
-        return result;
-    }
-
-    /**
-     * Filter values retrieved using tagList to include only those whose relator
-     * values are acceptable. Used for saving relators of authors separated by different
-     * types.
-     *
-     * @param record               The record (fed in automatically)
-     * @param tagList              The field specification to read
-     * @param acceptWithoutRelator Colon-delimited list of tags whose values should
-     * be accepted even if no relator subfield is defined
-     * @param relatorConfig        The setting in author-classification.ini which
-     * defines which relator terms are acceptable (or a colon-delimited list)
-     * @param acceptUnknownRelators Colon-delimited list of tags whose relators
-     * should be indexed even if they are not listed in author-classification.ini.
-     * @param indexRawRelators      Set to "true" to index relators raw, as found
-     * in the MARC or "false" to index mapped versions.
-     * @return List result
-     */
-    public List getRelatorsFilteredByRelator(Record record, String tagList,
-        String acceptWithoutRelator, String relatorConfig,
-        String acceptUnknownRelators, String indexRawRelators
-    ) {
-        // default firstOnly to false!
-        return getRelatorsFilteredByRelator(
-            record, tagList, acceptWithoutRelator, relatorConfig,
-            acceptUnknownRelators, indexRawRelators, false
-        );
-    }
-
-    /**
-     * Filter values retrieved using tagList to include only those whose relator
-     * values are acceptable. Used for saving relators of authors separated by different
-     * types.
-     *
-     * @param record               The record (fed in automatically)
-     * @param tagList              The field specification to read
-     * @param acceptWithoutRelator Colon-delimited list of tags whose values should
-     * be accepted even if no relator subfield is defined
-     * @param relatorConfig        The setting in author-classification.ini which
-     * defines which relator terms are acceptable (or a colon-delimited list)
-     * @param acceptUnknownRelators Colon-delimited list of tags whose relators
-     * should be indexed even if they are not listed in author-classification.ini.
-     * @return List result
-     */
-    public List getRelatorsFilteredByRelator(Record record, String tagList,
-        String acceptWithoutRelator, String relatorConfig,
-        String acceptUnknownRelators
-    ) {
-        // default firstOnly to false!
-        return getRelatorsFilteredByRelator(
-            record, tagList, acceptWithoutRelator, relatorConfig,
-            acceptUnknownRelators, "false", false
-        );
-    }
-
-    /**
-     * Filter values retrieved using tagList to include only those whose relator
-     * values are acceptable. Used for saving relators of authors separated by different
-     * types.
-     *
-     * @param record               The record (fed in automatically)
-     * @param tagList              The field specification to read
-     * @param acceptWithoutRelator Colon-delimited list of tags whose values should
-     * be accepted even if no relator subfield is defined
-     * @param relatorConfig        The setting in author-classification.ini which
-     * defines which relator terms are acceptable (or a colon-delimited list)
-     * @return List result
-     */
-    public List getRelatorsFilteredByRelator(Record record, String tagList,
-        String acceptWithoutRelator, String relatorConfig
-    ) {
-        // default firstOnly to false!
-        return getRelatorsFilteredByRelator(
-            record, tagList, acceptWithoutRelator, relatorConfig,
-            acceptWithoutRelator, "false", false
-        );
-    }
-
-    /**
-     * This method fetches relator definitions from ini file and casts them to an
-     * array. If a colon-delimited string is passed in, this will be directly parsed
-     * instead of resorting to .ini loading.
-     *
-     * @param setting Setting to load from .ini or colon-delimited list.
-     * @return String[]
-     */
-    protected String[] loadRelatorConfig(String setting){
-        StringBuilder relators = new StringBuilder();
-
-        // check for pipe-delimited string
-        String[] relatorSettings = setting.split("\\|");
-        for (String relatorSetting: relatorSettings) {
-            // check for colon-delimited string
-            String[] relatorArray = relatorSetting.split(":");
-            if (relatorArray.length > 1) {
-                for (int i = 0; i < relatorArray.length; i++) {
-                    relators.append(relatorArray[i]).append(",");
-                }
-            } else {
-                relators.append(getConfigSetting(
-                    "author-classification.ini", "AuthorRoles", relatorSetting
-                )).append(",");
-            }
-        }
-
-        return relators.toString().split(",");
-    }
-
-    /**
-     * Normalizes a relator string and returns a list containing the normalized
-     * relator plus any configured synonyms.
-     *
-     * @param relator Relator term to normalize
-     * @return List of strings
-     */
-    public List<String> normalizeRelatorAndAddSynonyms(String relator)
-    {
-        List<String> newList = new ArrayList<String>();
-        String normalized = normalizeRelatorString(relator);
-        newList.add(normalized);
-        String synonyms = getConfigSetting(
-            "author-classification.ini", "RelatorSynonyms", relator
-        );
-        if (null != synonyms && synonyms.length() > 0) {
-            for (String synonym: synonyms.split("\\|")) {
-                String normalizedSynonym = normalizeRelatorString(synonym);
-                relatorSynonymLookup.put(normalizedSynonym, relator);
-                newList.add(normalizedSynonym);
-            }
-        }
-        return newList;
-    }
-
-    /**
-     * Normalizes the strings in a list.
-     *
-     * @param stringList List of strings to be normalized
-     * @return Normalized List of strings 
-     */
-    protected List<String> normalizeRelatorStringList(List<String> stringList)
-    {
-        List<String> newList = new ArrayList<String>();
-        for (String relator: stringList) {
-            newList.addAll(normalizeRelatorAndAddSynonyms(relator));
-        }
-        return newList;
-    }
-
-    /**
-     * Normalizes a string
-     *
-     * @param string String to be normalized
-     * @return string
-     */
-    protected String normalizeRelatorString(String string)
-    {
-        return string
-            .trim()
-            .toLowerCase()
-            .replaceAll("\\p{Punct}+", "");    //POSIX character class Punctuation: One of !"#$%&'()*+,-./:;<=>?@[\]^_`{|}~
-    }
-
-    /**
-     * Filter values retrieved using tagList to include only those whose relator
-     * values are acceptable. Used for separating different types of authors.
-     *
-     * @param record               The record (fed in automatically)
-     * @param tagList              The field specification to read
-     * @param acceptWithoutRelator Colon-delimited list of tags whose values should
-     * be accepted even if no relator subfield is defined
-     * @param relatorConfig        The setting in author-classification.ini which
-     * defines which relator terms are acceptable (or a colon-delimited list)
-     * @param acceptUnknownRelators Colon-delimited list of tags whose relators
-     * should be indexed even if they are not listed in author-classification.ini.
-     * @param indexRawRelators      Set to "true" to index relators raw, as found
-     * in the MARC or "false" to index mapped versions.
-     * @return List result
-     */
-    public List<String> getAuthorInitialsFilteredByRelator(Record record,
-        String tagList, String acceptWithoutRelator, String relatorConfig,
-        String acceptUnknownRelators, String indexRawRelators
-    ) {
-        List<String> authors = getAuthorsFilteredByRelator(
-            record, tagList, acceptWithoutRelator, relatorConfig,
-            acceptUnknownRelators, indexRawRelators
-        );
-        List<String> result = new LinkedList<String>();
-        for (String author : authors) {
-            result.add(processInitials(author));
-        }
-        return result;
-    }
-
-    /**
-     * Filter values retrieved using tagList to include only those whose relator
-     * values are acceptable. Used for separating different types of authors.
-     *
-     * @param record               The record (fed in automatically)
-     * @param tagList              The field specification to read
-     * @param acceptWithoutRelator Colon-delimited list of tags whose values should
-     * be accepted even if no relator subfield is defined
-     * @param relatorConfig        The setting in author-classification.ini which
-     * defines which relator terms are acceptable (or a colon-delimited list)
-     * @return List result
-     */
-    public List<String> getAuthorInitialsFilteredByRelator(Record record,
-        String tagList, String acceptWithoutRelator, String relatorConfig
-    ) {
-        return getAuthorInitialsFilteredByRelator(
-            record, tagList, acceptWithoutRelator, relatorConfig,
-            acceptWithoutRelator, "false"
-        );
-    }
-
-    /**
-     * Filter values retrieved using tagList to include only those whose relator
-     * values are acceptable. Used for separating different types of authors.
-     *
-     * @param record               The record (fed in automatically)
-     * @param tagList              The field specification to read
-     * @param acceptWithoutRelator Colon-delimited list of tags whose values should
-     * be accepted even if no relator subfield is defined
-     * @param relatorConfig        The setting in author-classification.ini which
-     * defines which relator terms are acceptable (or a colon-delimited list)
-     * @param acceptUnknownRelators Colon-delimited list of tags whose relators
-     * should be indexed even if they are not listed in author-classification.ini.
-     * @return List result
-     */
-    public List<String> getAuthorInitialsFilteredByRelator(Record record,
-        String tagList, String acceptWithoutRelator, String relatorConfig,
-        String acceptUnknownRelators
-    ) {
-        return getAuthorInitialsFilteredByRelator(
-            record, tagList, acceptWithoutRelator, relatorConfig,
-            acceptUnknownRelators, "false"
-        );
-    }
-    
-    /**
-     * Takes a name and cuts it into initials
-     * @param authorName e.g. Yeats, William Butler
-     * @return initials e.g. w b y wb
-     */
-    protected String processInitials(String authorName) {
-        Boolean isPersonalName = false;
-        // we guess that if there is a comma before the end - this is a personal name
-        if ((authorName.indexOf(',') > 0) 
-            && (authorName.indexOf(',') < authorName.length()-1)) {
-            isPersonalName = true;
-        }
-        // get rid of non-alphabet chars but keep hyphens and accents 
-        authorName = authorName.replaceAll("[^\\p{L} -]", "").toLowerCase();
-        String[] names = authorName.split(" "); //split into tokens on spaces
-        // if this is a personal name we'll reorganise to put lastname at the end
-        String result = "";
-        if (isPersonalName) {
-            String lastName = names[0]; 
-            for (int i = 0; i < names.length-1; i++) {
-                names[i] = names[i+1];
-            }
-            names[names.length-1] = lastName;
-        }
-        // put all the initials together in a space separated string
-        for (String name : names) {
-            if (name.length() > 0) {
-                String initial = name.substring(0,1);
-                // if there is a hyphenated name, use both initials
-                int pos = name.indexOf('-');
-                if (pos > 0 && pos < name.length() - 1) {
-                    String extra = name.substring(pos+1, pos+2);
-                    initial = initial + " " + extra;
-                }
-                result += " " + initial; 
-            }
-        }
-        // grab all initials and stick them together
-        String smushAll = result.replaceAll(" ", "");
-        // if it's a long personal name, get all but the last initials as well
-        // e.g. wb for william butler yeats
-        if (names.length > 2 && isPersonalName) {
-            String smushPers = result.substring(0,result.length()-1).replaceAll(" ","");
-            result = result + " " + smushPers;
-        }
-        // now we have initials separate and together
-        if (!result.trim().equals(smushAll)) {
-            result += " " + smushAll; 
-        }
-        result = result.trim();
-        return result;
     }
 
     /**
